refactor: move Location class to separate module
